requirements.txt
django-browser-reload==1.18.0
django-cotton==2.0.0
django-environ==0.12.0
django-tailwind==3.8.0
gunicorn==23.0.0
pillow==11.1.0
pipdeptree==2.25.1
psycopg2-binary==2.9.10
supabase==2.15.0
whitenoise==6.9.0
